+++ src/index.d.ts
@@ -5,6 +5,9 @@ import {
     v2_0 as fspiopAPI,
     thirdparty as tpAPI
 } from '@mojaloop/api-snippets'
+
+import * as ilp from './ilp'
+
 declare namespace SDKStandardComponents {
 
     /* hashmap of versions of various resources */
@@ -104,7 +107,7 @@ declare namespace SDKStandardComponents {
          */
         putConsentsError(
             consentId: string,
-            consentBody: fspiopAPI.Schemas.ErrorInformationObject,
+            consentBody: fspiopAPI.Types.ErrorInformationObject,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -167,7 +170,7 @@ declare namespace SDKStandardComponents {
          */
         putConsentRequestsError(
             consentRequestId: string,
-            consentRequestBody: fspiopAPI.Schemas.ErrorInformationObject,
+            consentRequestBody: fspiopAPI.Types.ErrorInformationObject,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -237,7 +240,7 @@ declare namespace SDKStandardComponents {
          * @returns {Promise<object>} JSON response body if one was received
          */
         putThirdpartyRequestsTransactionsError(
-            thirdpartyRequestsTransactionsBody: fspiopAPI.Schemas.ErrorInformationObject,
+            thirdpartyRequestsTransactionsBody: fspiopAPI.Types.ErrorInformationObject,
             transactionRequestId: string,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
@@ -318,7 +321,7 @@ declare namespace SDKStandardComponents {
          */
         putAccountsError (
             userId: string,
-            accountsBody: fspiopAPI.Schemas.ErrorInformationObject,
+            accountsBody: fspiopAPI.Types.ErrorInformationObject,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -353,7 +356,7 @@ declare namespace SDKStandardComponents {
          */
          putServicesError (
             serviceType: string,
-            servicesBody: fspiopAPI.Schemas.ErrorInformationObject,
+            servicesBody: fspiopAPI.Types.ErrorInformationObject,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -395,7 +398,7 @@ declare namespace SDKStandardComponents {
          * @returns {Promise<object>} JSON response body if one was received
          */
         putThirdpartyRequestsVerificationsError(
-            thirdpartyRequestsVerificationsBody: fspiopAPI.Schemas.ErrorInformationObject,
+            thirdpartyRequestsVerificationsBody: fspiopAPI.Types.ErrorInformationObject,
             verificationsRequestId: string,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
@@ -431,7 +434,7 @@ declare namespace SDKStandardComponents {
             idType: string,
             idValue: string,
             idSubValue: string | undefined,
-            body: fspiopAPI.Schemas.Party,
+            body: fspiopAPI.Types.Party,
             destFspId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -449,7 +452,7 @@ declare namespace SDKStandardComponents {
             idType: string,
             idValue: string,
             idSubValue: string | undefined,
-            error: fspiopAPI.Schemas.ErrorInformationObject,
+            error: fspiopAPI.Types.ErrorInformationObject,
             destFspId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -462,7 +465,7 @@ declare namespace SDKStandardComponents {
          * @returns {Promise<object>} JSON response body if one was received
          */
         postQuotes(
-            quoteRequest: fspiopAPI.Schemas.QuotesPostRequest,
+            quoteRequest: fspiopAPI.Types.QuotesPostRequest,
             destParticipantId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>;
 
@@ -476,7 +479,7 @@ declare namespace SDKStandardComponents {
          */
         putAuthorizations(
             transactionRequestId: string,
-            authorizationResponse: fspiopAPI.Schemas.AuthorizationsIDPutResponse,
+            authorizationResponse: fspiopAPI.Types.AuthorizationsIDPutResponse,
             destFspId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>
 
@@ -490,7 +493,7 @@ declare namespace SDKStandardComponents {
          */
         putAuthorizationsError(
             transactionRequestId: string,
-            error: fspiopAPI.Schemas.ErrorInformationObject,
+            error: fspiopAPI.Types.ErrorInformationObject,
             destFspId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>
 
@@ -502,7 +505,7 @@ declare namespace SDKStandardComponents {
          * @param {string} destFspId The id of the destination participant
          */
         postParticipants(
-            participantRequest: fspiopAPI.Schemas.ParticipantsPostRequest,
+            participantRequest: fspiopAPI.Types.ParticipantsPostRequest,
             destFspId: string
         ): Promise<GenericRequestResponse | GenericRequestResponseUndefined>
     }
@@ -735,6 +738,7 @@ declare namespace SDKStandardComponents {
             PAYEE_REJECTED_TXN:               MojaloopApiErrorCode
             PAYEE_FSP_REJECTED_TXN:           MojaloopApiErrorCode
             PAYEE_UNSUPPORTED_CURRENCY:       MojaloopApiErrorCode
+            PAYEE_IDENTIFIER_NOT_VALID:       MojaloopApiErrorCode
             PAYEE_LIMIT_ERROR:                MojaloopApiErrorCode
             PAYEE_PERMISSION_ERROR:           MojaloopApiErrorCode
             GENERIC_PAYEE_BLOCKED_ERROR:      MojaloopApiErrorCode
@@ -864,6 +868,33 @@ declare namespace SDKStandardComponents {
 
         var validator: typeof JwsValidator
         var signer: typeof JwsSigner
+    }
+
+    interface IlpProcessor {
+        new (options: IlpOptions): IlpProcessor;
+        version: keyof typeof Ilp.ILP_VERSIONS;
+        calculateConditionFromFulfil(fulfilment: string): string;
+        calculateFulfil(transactionObject: ilp.TransactionObject | string): string
+        decodeIlpPacket(ilpPacket: string): ilp.IlpInputV1 | ilp.IlpInputV4;
+        getFxQuoteResponseIlp(fxQuoteRequest: ilp.FxQuoteRequest, fxQuoteBeResponse: ilp.FxQuoteBeResponse): ilp.IlpResponse;
+        getResponseIlp(transactionObject: ilp.TransactionObject): ilp.IlpResponse;
+        getTransactionObject(ilpPacket: string): ilp.TransactionObject;
+        validateFulfil(fulfilment: string, condition: string): boolean
+        // add other public methods and define needed types
+    }
+
+    type IlpOptions = {
+        secret: string;
+        logger: Logger.Logger;
+    }
+
+    namespace Ilp {
+        const ilpFactory: (version: keyof typeof ILP_VERSIONS, options: IlpOptions) => IlpProcessor;
+
+        const ILP_VERSIONS: Readonly<{
+            v1: 'v1';
+            v4: 'v4';
+        }>;
     }
 }